requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

streamlit==1.44.1
timm==1.0.15
huggingface_hub==0.30.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
Pillow==11.0.0